Python/Django/requirement.txt
django
psycopg2-binary
python-dotenv
whitenoise
djangorestframework
pygments
django-extensions
ipython